Rename workspaces plugin file to match other plugins, clean up exports
