+++ packages/apputils-extension/src/index.ts
@@ -21,10 +21,6 @@ import {
   sessionContextDialogs
 } from '@jupyterlab/apputils';
 
-import { IFileBrowserFactory } from '@jupyterlab/filebrowser';
-
-import { IMainMenu } from '@jupyterlab/mainmenu';
-
 import { URLExt } from '@jupyterlab/coreutils';
 
 import { IStateDB, StateDB } from '@jupyterlab/statedb';
@@ -41,9 +37,9 @@ import { Palette } from './palette';
 
 import { settingsPlugin } from './settingsplugin';
 
-import { themesPlugin, themesPaletteMenuPlugin } from './themeplugins';
+import { themesPlugin, themesPaletteMenuPlugin } from './themesplugins';
 
-import { WorkspaceUI } from './workspace-ui';
+import { workspacesPlugin } from './workspacesplugin';
 
 /**
  * The interval in milliseconds before recover options appear during splash.
@@ -494,16 +490,6 @@ const utilityCommands: JupyterFrontEndPlugin<void> = {
 };
 
 /**
- * The workspace UI extension.
- */
-const workspaceUI: JupyterFrontEndPlugin<void> = {
-  activate: WorkspaceUI.activate,
-  id: '@jupyterlab/apputils-extension:workspace-ui',
-  autoStart: true,
-  requires: [IMainMenu, IFileBrowserFactory, IWindowResolver, IStateDB, IRouter]
-};
-
-/**
  * Export the plugins as default.
  */
 const plugins: JupyterFrontEndPlugin<any>[] = [
@@ -518,7 +504,7 @@ const plugins: JupyterFrontEndPlugin<any>[] = [
   themesPlugin,
   themesPaletteMenuPlugin,
   utilityCommands,
-  workspaceUI
+  workspacesPlugin
 ];
 export default plugins;
